pyproject.toml
[tool.poetry]
name = "1pic1day"
version = "0.1.0"
description = "Upload an album photo where only one picture per day can be seen"
authors = ["MattiooFR <13595614+MattiooFR@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask_script = "^2.0.6"
flask_migrate = "^2.5.3"
psycopg2-binary = "^2.8.5"
flask_cors = "^3.0.8"
gunicorn = "^20.0.4"
requests = "^2.23.0"
authlib = "^0.14.1"
six = "^1.14.0"
flask-wtf = "^0.14.3"
flask-uploads = {git = "https://github.com/maxcountryman/flask-uploads.git"}
werkzeug = "^1.0.1"
python-jose = "^3.1.0"
bootstrap-flask = "^1.3.1"
flask-testing = "^0.8.0"
boto3 = "^1.14.15"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.13.0"
flake8-awesome = "^1.0"
black = {version = "^18.3-alpha.0", allow-prereleases = true}
pytest = "^5.0"
pytest-cov = "^2.7"
Flask-Testing = "^0.8.0"
flake8 = "^3.8.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
